Refaktorointi ja Android-korjaukset
