+++ .gitignore
@@ -12,5 +12,5 @@ web-build/
 
 # macOS
 .DS_Store
-firebaseConfig.js
 *.csv
+.env
