+++ package.json
@@ -30,6 +30,7 @@
     "jest-expo": "^48.0.2",
     "react": "18.2.0",
     "react-native": "0.71.4",
+    "react-native-dotenv": "^3.4.8",
     "react-native-event-listeners": "^1.0.7",
     "react-native-geocoder": "^0.5.0",
     "react-native-gesture-handler": "~2.9.0",
