+++ src/components/LoggedIn.js
@@ -1,0 +1,27 @@
+import { signOut } from "firebase/auth";
+import { View, Text, Button } from "react-native";
+import { auth } from "./firebaseConfig";
+
+export default function LoggedIn() {
+  const logout = async () => {
+    try {
+      await signOut(auth);
+    } catch (e) {
+      console.error(e);
+    }
+  };
+  const user = auth.currentUser;
+
+  return (
+    <View>{user ? (
+      <>
+        <Text>Email:{user.email}</Text>
+        <Text>Username:{user.displayName}</Text>
+      </>
+    ) : (
+      <Text></Text>
+    )}
+      <Button title="Log out" onPress={logout} />
+    </View>
+  );
+}
