+++ src/components/Signup.js
@@ -2,40 +2,40 @@ import { View, Text, Button, TextInput, StyleSheet, Alert } from "react-native";
 import { createUserWithEmailAndPassword, signInWithEmailAndPassword, updateProfile } from "firebase/auth";
 import { useNavigation } from "@react-navigation/native";
 import React, { useState } from "react";
-import { auth } from "../firebaseConfig";
+import { auth } from "./firebaseConfig";
 
 export default function Signup({ navigation }) {
-    const [username, setUsername] = useState('')
-    const [email, setEmail] = useState('');
-    const [password, setPassword] = useState('');
-    const [confirmPassword, setConfirmPassword] = useState('');
-    const [error, setError] = useState(null);
-  
-    const createAccount = async () => {
-        try {
-            if (password === confirmPassword) {
-                let res = await createUserWithEmailAndPassword(auth, email, password);
-                if (res && res.user) {
-                  Alert.alert("Käyttäjän luominen onnistui.");
-                  await signInWithEmailAndPassword(auth, email, password); // Sign in the user to get the updated currentUser object
-                  updateProfile(auth.currentUser, {
-                    displayName: username
-                  }).then(() => {
-                    navigation.replace("Login");
-                  }).catch((e) => {
-                    console.log(e);
-                  })
-                }
-            } else {
-                setError('Salasanat eivät täsmää')
-            }
-        } catch (e) {
-            setError('Käyttäjän luominen epäonnistui');
-            console.log(e);
-        }
-    };
+  const [username, setUsername] = useState('')
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [error, setError] = useState(null);
 
-    return (
+  const createAccount = async () => {
+    try {
+      if (password === confirmPassword) {
+        let res = await createUserWithEmailAndPassword(auth, email, password);
+        if (res && res.user) {
+          Alert.alert("Käyttäjän luominen onnistui.");
+          await signInWithEmailAndPassword(auth, email, password); // Sign in the user to get the updated currentUser object
+          updateProfile(auth.currentUser, {
+            displayName: username
+          }).then(() => {
+            navigation.replace("Login");
+          }).catch((e) => {
+            console.log(e);
+          })
+        }
+      } else {
+        setError('Salasanat eivät täsmää')
+      }
+    } catch (e) {
+      setError('Käyttäjän luominen epäonnistui');
+      console.log(e);
+    }
+  };
+
+  return (
     <View style={styles.outer}>
       <View style={styles.inner}>
         <Text style={styles.header}>Signup</Text>
@@ -85,33 +85,33 @@ export default function Signup({ navigation }) {
         />
       </View>
     </View>
-    );
+  );
 }
 
 const styles = StyleSheet.create({
-    outer: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center',
-    },
-    inner: {
-      width: 240,
-    },
-    header: {
-      fontSize: 24,
-      fontWeight: 'bold',
-      marginBottom: 20,
-    },
-    input: {
-      borderWidth: 1,
-      borderColor: '#ccc',
-      borderRadius: 4,
-      paddingVertical: 8,
-      paddingHorizontal: 12,
-      marginBottom: 16,
-    },
-    error: {
-      marginBottom: 20,
-      color: 'red',
-    },
-  });
+  outer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  inner: {
+    width: 240,
+  },
+  header: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginBottom: 20,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    borderRadius: 4,
+    paddingVertical: 8,
+    paddingHorizontal: 12,
+    marginBottom: 16,
+  },
+  error: {
+    marginBottom: 20,
+    color: 'red',
+  },
+});
